pyproject.toml
[project]
name = "anthropic-mcp-connector"
version = "0.1.0"
description = "Test for Anthropic MCP Connector"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.52.1",
    "python-dotenv",
    "pytz",
    "requests"
]
